Remove padding for last column
For very long columns (such as cmdline in jps) the formatting looks
pretty goofy. The divider line is still very unpleasant to look at, but
removing the padding on the last column at least avoids every line
having a bunch of whitespace at the end and introducing unnecessary line
breaks.